# Support group-limit optimization for `ROW_NUMBER`