@mythrocks 2024-12-18
> Just curious if you ran any performance numbers on this?  

I'm afraid I did run a couple of tests. 

Before the optimization:
```scala
scala> spark.time( sql(" select * from ( select *, row_number() over (partition by gby order by oby) rn from parquet.`/home/mithunr/workspace/Demos/rownumber-grouplimits/test_repart_10/` order by 1,2,3 ) where rn < 4 ").show )
...
      !Exec <WindowGroupLimitExec> cannot run on GPU because Only Rank() and DenseRank() are currently supported for window group limits
...

+---+---+---+
|gby|oby| rn|
+---+---+---+
|  1|  1|  1|
|  1|  1|  2|
|  1|  1|  3|
|  2|  1|  1|
|  2|  1|  2|
|  2|  1|  3|
|  5|  1|  1|
|  5|  1|  2|
|  5|  1|  3|
+---+---+---+
Time taken: 6098 ms
```

After this change:
```scala
scala> spark.time( sql(" select * from ( select *, row_number() over (partition by gby order by oby) rn from parquet.`/home/mithunr/workspace/Demos/rownumber-grouplimits/test_repart_10/` order by 1,2,3 ) where rn < 4 ").show )
...
      *Expression <Alias> row_number() windowspecdefinition(gby#55, oby#56 ASC NULLS FIRST, specifiedwindowframe(RowFrame, unboundedpreceding$(), currentrow$())) AS rn#53 will run on GPU
...
          *Expression <WindowSpecDefinition> windowspecdefinition(gby#55, oby#56 ASC NULLS FIRST, specifiedwindowframe(RowFrame, unboundedpreceding$(), currentrow$())) will run on GPU
...
+---+---+---+
|gby|oby| rn|
+---+---+---+
|  1|  1|  1|
|  1|  1|  2|
|  1|  1|  3|
|  2|  1|  1|
|  2|  1|  2|
|  2|  1|  3|
|  5|  1|  1|
|  5|  1|  2|
|  5|  1|  3|
+---+---+---+

Time taken: 2959 ms
```

So about 2x faster for `row_number`, for a 200M row dataset with only 3 group-by keys (i.e. `gby`).  Not comprehensive, exactly, but I'm cautiously optimistic.